create a shared reparent-to-flex apply function, since the two functions were 97% identical
