+++ editor/src/components/canvas/canvas-strategies/absolute-reparent-to-flex-strategy.tsx
@@ -27,17 +27,13 @@ import { updateHighlightedViews } from '../commands/update-highlighted-views-com
 import { CanvasCommand, foldAndApplyCommandsInner } from '../commands/commands'
 import { deleteProperties } from '../commands/delete-properties-command'
 import { updateSelectedViews } from '../commands/update-selected-views-command'
-import { findReparentStrategy, getReparentTargetForFlexElement } from './reparent-strategy-helpers'
+import {
+  applyFlexReparent,
+  findReparentStrategy,
+  getReparentTargetForFlexElement,
+} from './reparent-strategy-helpers'
 import { getReparentCommands } from './reparent-utils'
 import { ifAllowedToReparent } from './reparent-helpers'
-
-const propertiesToRemove: Array<PropertyPath> = [
-  PP.create(['style', 'position']),
-  PP.create(['style', 'left']),
-  PP.create(['style', 'top']),
-  PP.create(['style', 'right']),
-  PP.create(['style', 'bottom']),
-]
 
 export const absoluteReparentToFlexStrategy: CanvasStrategy = {
   id: 'ABSOLUTE_REPARENT_TO_FLEX',
@@ -104,87 +100,6 @@ export const absoluteReparentToFlexStrategy: CanvasStrategy = {
     interactionSession: InteractionSession,
     strategyState: StrategyState,
   ): StrategyApplicationResult {
-    const filteredSelectedElements = getDragTargets(canvasState.selectedElements)
-
-    return ifAllowedToReparent(canvasState, strategyState, filteredSelectedElements, () => {
-      if (
-        interactionSession.interactionData.type == 'DRAG' &&
-        interactionSession.interactionData.drag != null
-      ) {
-        const reparentResult = getReparentTargetForFlexElement(
-          filteredSelectedElements,
-          interactionSession,
-          canvasState,
-          strategyState,
-        )
-
-        if (
-          reparentResult.shouldReparent &&
-          reparentResult.newParent != null &&
-          filteredSelectedElements.length === 1
-        ) {
-          const target = filteredSelectedElements[0]
-          const newParent = reparentResult.newParent
-          // Reparent the element.
-          const newPath = EP.appendToPath(reparentResult.newParent, EP.toUid(target))
-          const reparentCommands = getReparentCommands(
-            canvasState.builtInDependencies,
-            canvasState.projectContents,
-            canvasState.nodeModules,
-            canvasState.openFile,
-            target,
-            reparentResult.newParent,
-          )
-
-          // Strip the `position`, positional and dimension properties.
-          const commandToRemoveProperties = deleteProperties('always', newPath, propertiesToRemove)
-
-          const commandsBeforeReorder = [
-            ...reparentCommands,
-            updateSelectedViews('always', [newPath]),
-          ]
-
-          const commandsAfterReorder = [
-            commandToRemoveProperties,
-            setElementsToRerenderCommand([newPath]),
-            updateHighlightedViews('mid-interaction', []),
-            setCursorCommand('mid-interaction', CSSCursor.Move),
-          ]
-
-          let commands: Array<CanvasCommand>
-          if (reparentResult.shouldReorder) {
-            // Reorder the newly reparented element into the flex ordering.
-            const pointOnCanvas = offsetPoint(
-              interactionSession.interactionData.dragStart,
-              interactionSession.interactionData.drag,
-            )
-
-            const siblingsOfTarget = MetadataUtils.getChildrenPaths(
-              strategyState.startingMetadata,
-              newParent,
-            )
-
-            const newIndex = getReorderIndex(
-              strategyState.startingMetadata,
-              siblingsOfTarget,
-              pointOnCanvas,
-            )
-            commands = [
-              ...commandsBeforeReorder,
-              reorderElement('always', newPath, newIndex),
-              ...commandsAfterReorder,
-            ]
-          } else {
-            commands = [...commandsBeforeReorder, ...commandsAfterReorder]
-          }
-
-          return {
-            commands: commands,
-            customState: strategyState.customStrategyState,
-          }
-        }
-      }
-      return emptyStrategyApplicationResult
-    })
+    return applyFlexReparent('strip-absolute-props', canvasState, interactionSession, strategyState)
   },
 }
